Cleaned up header files and segue.
The ‘back’ button that used to appear when you logged out and in again
is gone. I made the code respond to a button tap instead of linking it
to a segue. All methods are now declared in the .h files, although we
should add more comments I expect, and some may be unnecessary.