Starting to check for user's game status.
Added a new class of view controller for if the user is in a game that
hasn’t started yet. Every time the app opens or a new user logs in, it
checks whether they are in a game and prints a message. (We will need
to change the behavior when returning from create or join game.)