Fixed save vs. saveInBackground vs. saveInBackgroundWithBlock
Now there should be no saves, and uses blocks if the code needs to wait
for save before segueing, or doing some action that requires the
objectId of a new object. Also, unwinds are called programmatically
within block in buttonTapped rather than having an additional
storyboard segue. Finally, if a user leaves an unstarted game, the
numberOfPlayers should and now does decrease as well.